feat: dont error when the helm directory is not found (#102)
Not error out will have a better behavior. The lib retrieves the
directory from the ArgoCD app in argocd, which is controlled by
definitions defined in `coopnorge/kubernetes-projects`

If there is a directory configured which does not exist due to a mistake
or due to it not yet being on the main branch then it is blocking if we
error here.
If a user makes a mistake on configuring the directory then the only
harm is that there are no manifests synced. We prefer to not block in
this case.